Split Makefiles in two, restore building of lib
I realised that the Python interface would require a library file, so
restored the Linux makefile to build the library.

Split the Windows and Linux makefiles into two, sharing a common import,
since it was getting overly complicated. Hopefully I can instruct Python
to use a different Makefile for each.