Build a wrapper around PyGraph, reimplement
The original PyBliss had two levels of wrappers, one (intpybliss) which
made the C++ functions accessible to Python, and another (PyBliss) which
was a full and independent Python Graph class, which called intpyblss
when it needed to do the heavy lifting like finding automorphisms.

I thought I could avoid the second by implementing all the features
straight into the C++ wrapper, but it turns out that PyBliss adds
features which aren't in the underlying C++ implementation (so far I've
seen node and edge deletion), and can't be easily added.

Reimplemented everything which I'd built so far, but haven't extended it
yet.